Changed proect structure
