+++ Vacations/ClientApp/src/app/app.module.ts
@@ -6,7 +6,6 @@ import { HTTP_INTERCEPTORS } from '@angular/common/http';
 import { HttpModule } from '@angular/http';
 
 import { AppComponent } from './app.component';
-import { ProfileModule } from './profile/profile.module';
 import { AuthModule } from './auth/auth.module';
 import { AppRoutes } from './app.routes';
 import { MyFirstInterceptor, MyDataService } from './app.service';
@@ -25,7 +24,6 @@ import { MainModule } from './main/main.module';
     RouterModule.forRoot(AppRoutes),
 
     MainModule,
-    ProfileModule,
     AuthModule
   ],
   providers: [{
