--- Vacations/ClientApp/src/app/main/components/my-vacations/my-vacations.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-my-vacations',
-  templateUrl: './my-vacations.component.html',
-  styleUrls: ['./my-vacations.component.scss']
-})
-export class MyVacationsComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
