+++ Vacations/ClientApp/src/app/main/components/profile/banner/banner.component.scss
@@ -5,7 +5,7 @@
     border-radius: 10px;
     border: 1px solid #878787;
     font-family: Muli-Bold;
-    background-image: url('../../../../assets/home-request-background.png');
+    background-image: url('../../../../../assets/home-request-background.png');
     background-repeat: no-repeat;
     background-size: cover;
