+++ Vacations/ClientApp/src/app/main/components/profile/banner/banner.component.ts
