+++ Vacations/ClientApp/src/app/main/components/profile/my-vacations/my-vacations.component.ts
@@ -1,14 +1,15 @@
 import { Component, OnInit } from '@angular/core';
 
-import { VacationService } from '../../services/vacation.service';
 import { Vacation } from './vacation.model';
+import { VacationService } from '../../../services/vacation.service';
 
 @Component({
-  selector: 'app-vacation',
-  templateUrl: './vacation.component.html',
-  styleUrls: ['./vacation.component.css']
+  selector: 'app-my-vacations',
+  templateUrl: './my-vacations.component.html',
+  styleUrls: ['./my-vacations.component.scss']
 })
-export class VacationComponent implements OnInit {
+export class MyVacationsComponent implements OnInit {
+
   vacations: Vacation[];
 
   constructor(private service: VacationService) { }
@@ -17,5 +18,5 @@ export class VacationComponent implements OnInit {
     this.service.getVacations()
       .subscribe(response => this.vacations = response);
   }
-
 }
+
