+++ Vacations/ClientApp/src/app/main/components/profile/profile.component.ts
@@ -1,23 +1,13 @@
 import { Component, OnInit } from '@angular/core';
 
-import { EmployeeService } from '../../services/employee.service';
-import { Employee } from './employee.model';
-
 @Component({
   selector: 'app-profile',
-  templateUrl: './profile.component.html',
-  styleUrls: ['./profile.component.scss']
+  templateUrl: './profile.component.html'
 })
 
 export class ProfileComponent implements OnInit {
-  title = 'profile';
-  employee: Employee;
-  
-  constructor(private service: EmployeeService) { }
+  constructor() { }
 
   ngOnInit() {
-    this.service.getEmployee().subscribe(response => this.employee = response);
   }
-
-
 }
