--- Vacations/ClientApp/src/app/main/components/vacation/vacation.component.html
@@ -1,12 +1,0 @@
-<div >
-    <h3>Vacations</h3>
-
-
-    <div class="details">
-        <form>
-            <div *ngFor="let vac of vacations">
-                {{ vac.startVocationDate }} <br>
-            </div>
-        </form>
-    </div>
-</div>
