+++ Vacations/ClientApp/src/app/main/main.component.html
@@ -5,16 +5,10 @@
     <div class="col-2">
       <app-left-navigation></app-left-navigation>
     </div>
-    <div class="col-3 mt25">
-      <app-profile></app-profile>
-    </div>
-    <div class="col-5 mt25">
-      <app-banner></app-banner>
-      <div class="mt25">
-        <!-- <app-vacation-request></app-vacation-request> -->
-        <app-my-vacations></app-my-vacations>
-      </div>
-    </div>
+    <router-outlet></router-outlet>
     <div class="col-2"></div>
   </div>
 </div>
+
+    
+
