+++ Vacations/ClientApp/src/app/main/services/employee.service.ts
@@ -2,7 +2,7 @@ import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs/Observable';
 
-import { Employee } from '../components/profile/employee.model';
+import { Employee } from '../components/profile/my-profile/employee.model';
 import { environment } from '../../../environments/environment';
 
 @Injectable()
