--- Vacations/ClientApp/src/app/profile/counter/counter.component.html
@@ -1,7 +1,0 @@
-<h1>Counter</h1>
-
-<p>This is a simple example of an Angular component.</p>
-
-<p>Current count: <strong>{{ currentCount }}</strong></p>
-
-<button (click)="incrementCounter()">Increment</button>
