--- Vacations/ClientApp/src/app/profile/employee-profile/employee-profile.component.css
@@ -1,11 +1,0 @@
-.profileBorder {
-    border: 0.5px solid grey;
-    color: silver;
-} 
-
-.border {
-    border-radius: 15px;
-    background-color: #29292c;
-    border: 1px solid grey;
-    color: silver;
-}
