--- Vacations/ClientApp/src/app/profile/employee-profile/employee-profile.component.html
@@ -1,64 +1,0 @@
-  <div class="row">
-    <div class="col-sm-4">
-        <div class="border" >
-      <div >
-        <h1> My Profile </h1>
-      </div>
-
-      <div  *ngIf="employee; else loading">
-        <div class="profileBorder">
-          <p>
-            <b> {{ employee.name }} </b>   
-            <b>{{employee.surname}} </b>
-            <p *ngIf="employee.jobTitle; else loading"> <br>  {{employee.jobTitle}}  
-          </p>
-        </div>
-        <div >
-          <h6> date of birth
-            <br> </h6>
-          <p *ngIf="employee.birthday; else loading">
-            {{ employee.birthday }}
-            <hr>
-
-            <h6> personal email
-              <br> </h6>
-            <p *ngIf="employee.email; else loading">
-              {{ employee.workEmail}}
-              <hr>
-
-              <h6> work email
-                <br> </h6>
-              <p *ngIf="employee.workEmail; else loading">
-                {{ employee.workEmail}}
-                <hr>
-
-                <h6> phone number
-                  <br> </h6>
-                <p *ngIf="employee.telephoneNumber; else loading">
-                  +380 {{employee.telephoneNumber}}
-                  <hr>
-
-                  <h6> skype
-                    <br> </h6>
-                  <p *ngIf="employee.skype; else loading">
-                    {{employee.skype}}
-                    <hr>
-
-                    <h6> hire date
-                      <br> </h6>
-                    <p *ngIf="employee.startDate; else loading">
-                      {{employee.startDate}} - {{employee.endDate}}
-                      <hr>
-
-                      <h6> team name
-                        <br> </h6>
-                      <p *ngIf="employee.team; else loading">
-                        {{employee.team}} <br>
-                      </p>
-        </div>
-      </div>
-
-    </div>
-  </div>
-</div>
-<ng-template #loading>Loading...</ng-template>
