--- Vacations/ClientApp/src/app/profile/employee-profile/employee.model.ts
@@ -1,14 +1,0 @@
-export interface Employee {
-    name: string;
-    surname: string;
-    workEmail: string;
-    telephoneNumber: number;
-    birthday: Date;
-    skype: string;
-    startDate: Date;
-    endDate: Date;
-    balance: number;
-    employeeStatus: string;
-    jobTitle: string;
-    team: string[];
-}
