--- Vacations/ClientApp/src/app/profile/profile.component.css
@@ -1,5 +1,0 @@
-@media (max-width: 767px) {
-  .body-content {
-    padding-top: 50px;
-  }
-}
